src/frontend/app.py
import streamlit as st
import requests
import json
import time
import sys
import os

# プロジェクトルートディレクトリをパスに追加して src モジュールを解決できるようにする
sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../../')))

# 設定
st.set_page_config(
    page_title="Tax-Mate AutoPay Security Demo",
    page_icon="🛡️",
    layout="wide"
)

API_URL = "http://localhost:8000"

# --- Helper Functions ---
def reset_system():
    try:
        requests.post(f"{API_URL}/reset")
        st.toast("System Reset Successfully!", icon="✅")
        # セッションステートもクリア
        for key in list(st.session_state.keys()):
            del st.session_state[key]
    except Exception as e:
        st.error(f"Failed to reset: {e}")

def get_logs():
    try:
        res = requests.get(f"{API_URL}/logs")
        return res.json().get("logs", [])
    except:
        return []

def run_vulnerable():
    st.session_state['vulnerable_running'] = True
    try:
        res = requests.post(f"{API_URL}/run/vulnerable", json={"invoice_text": st.session_state.get('invoice_text')})
        return res.json()
    except Exception as e:
        st.error(f"Error: {e}")
        return {}
    finally:
        st.session_state['vulnerable_running'] = False

def start_secure():
    st.session_state['secure_running'] = True
    try:
        res = requests.post(f"{API_URL}/run/secure/start", json={"invoice_text": st.session_state.get('invoice_text')})
        try:
            data = res.json()
        except json.JSONDecodeError:
            st.error(f"Server Error (Status {res.status_code}): {res.text}")
            return
            
        if res.status_code != 200:
            st.error(f"API Error: {data.get('detail', 'Unknown error')}")
            return
            
        # 完了後の処理
        st.session_state['secure_status'] = data.get('status')
        st.session_state['secure_final_output'] = data.get('final_output')
        st.session_state['secure_thread_id'] = data.get('thread_id')
        st.session_state['secure_logs_before'] = get_logs() # ログ取得（あまり意味ないかもだが）
        
    except Exception as e:
        st.error(f"Error starting secure agent: {e}")
    finally:
        st.session_state['secure_running'] = False

# resume_secure removed


# --- UI Layout ---
st.title("Tax-Mate AutoPay: Security Demo 🛡️")

st.markdown("""
このデモアプリは、自律型AIエージェントに対する **Indirect Prompt Injection** 攻撃と、
それに対する **Human-in-the-loop (HITL)** 防御を比較検証するためのものです。
\n
**シナリオ:** 経理担当AIが請求書を読み取り、支払処理を行います。しかし、請求書には攻撃者による隠し命令が含まれています。
""")

# Data Preparation
from src.data.invoices import POISONED_INVOICE_TEXT
if 'invoice_text' not in st.session_state:
    st.session_state['invoice_text'] = POISONED_INVOICE_TEXT

st.markdown("### 📨 受信した請求書 (攻撃コード入り)")
st.info("以下の請求書テキストには、**「ハッカーの口座へ送金しろ」** という隠し命令が含まれています。")
st.code(st.session_state['invoice_text'], language="text")

# Tabs
tab1, tab2 = st.tabs(["🔴 脆弱なエージェント (攻撃デモ)", "🟢 堅牢なエージェント (防御デモ)"])

# --- TAB 1: Vulnerable ---
with tab1:
    st.markdown("### 🔴 脆弱なエージェント")
    st.markdown("""
    このエージェントは、LLMの自律性を最大限に尊重し、ツール利用の確認を行いません。
    **検証ポイント:** 攻撃者の隠し命令（AWS口座の書き換えと送金）がそのまま実行されるか確認してください。
    """)
    
    if st.button("🚀 請求書を処理 (脆弱モード)", key="run_vuln"):
        with st.spinner("エージェントが処理中..."):
            result = run_vulnerable()
            if result:
                # Check outcome
                time.sleep(1) # Wait for log sync
                latest_logs = get_logs()
                hacked = any("HACKER-999" in log or "MAINTENANCE-999" in log for log in latest_logs)
                
                if hacked:
                    st.error("❌ 【AIが騙されました！】", icon="🚨")
                    st.markdown("""
                    ### 😱 攻撃成功（防御失敗）
                    **AIは請求書内の隠し命令に従い、攻撃者の口座へ送金を実行してしまいました。**
                    
                    データの改ざんと不正送金が発生しています。これが Indirect Prompt Injection の脅威です。
                    """)
                    st.error(f"🚨 検出された不正操作: {latest_logs[-1]}")
                else:
                    st.warning("⚠️ 攻撃は失敗したか、安全フィルターによってブロックされました。")
                    st.write("**エージェントの応答:**")
                    st.write(result.get("final_output"))


# --- TAB 2: Secure ---
with tab2:
    st.markdown("### 🟢 堅牢なエージェント (LLM Guardrail)")
    st.markdown("""
    このエージェントは、**LLMガードレール**によって守られています。
    ツール実行前に別のセキュリティAIが監査を行い、不正な操作（Prompt Injectionなど）を自動的にブロックします。
    **検証ポイント:** 攻撃者の命令が自動的に検知され、ブロックされるか確認してください。
    """)
    
    if st.button("🛡️ 安全なプロセスを開始 (防御モード)", key="start_sec"):
         with st.spinner("エージェント実行中 & ガードレール監査中..."):
             start_secure()
    
    if st.session_state.get('secure_status') == 'completed':
        final_output = st.session_state.get('secure_final_output', "")
        
        # ガードレールによるブロック判定
        # agents.py で指定した警告メッセージが含まれているか確認
        if "【セキュリティ警告】" in final_output and "ブロックされました" in final_output:
             st.success("✅ 【防御成功！】", icon="🛡️")
             st.markdown("""
             ### 🔒 ガードレールが攻撃を無効化しました
             **セキュリティAIが不正なツール操作を検知し、実行をブロックしました。**
             人手を介さずとも、自動的に攻撃を防ぐことができました。
             """)
             st.error(f"🤖 **エージェントからの報告:**Link\n{final_output}")
             
        else:
             st.info("プロセスが完了しました（ブロックなし）。")
             st.write("**エージェントの応答:**")
             st.write(final_output)

