pyproject.toml
[project]
name = "tax-mate-autopay"
version = "0.1.0"
description = "Security Demonstration for Tax-Mate AutoPay"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "streamlit>=1.35.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.8",
    "langchain-google-genai>=1.0.0",
    "google-genai",
    "langgraph>=0.1.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "langchain-groq>=1.1.1",
]

[tool.uv]
dev-dependencies = []
